organise files and refactor
