--- edit.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  
- 
-  <head>
-    <a href="/options.html"><img src="images/skeleton.png" class="rounded-circle" style="max-width:9%" alt="...">
-  </a>
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-<link href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@1,300&family=Roboto+Mono:ital,wght@1,300&family=Shippori+Antique&family=Shippori+Antique+B1&display=swap" rel="stylesheet">
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-      crossorigin="anonymous"
-    />
-    <link rel="stylesheet" href="style.css" />
-    <h1 class="text-center p-1">Edit Key</h1>
-  </head>
-   <body class="p-3 mb-2 bg-light text-dark wrapper">
-    <form class="container">
-      <div class="mb-3">
-        <p for="title" class="form-label">Add Bit to Key</p>
-        <input
-          type="text"
-          class="form-control"
-          id="new_url"
-          placeholder="http://example.com"
-        />
-      </div>
-      <button type="button" class="btn-sm btn-secondary fw-bolder" id="keyEdit">Add</button>
-    </form>
-      <p>Delete Bits</p>
-    <div class="container" id="allBits"></div>
-    
-    <div class="container mt-2">
-      <p>Delete key</p>
-      <button type="button" id="deleteKey" class="btn-sm btn-secondary fw-bolder">Delete</button>
-    </div>
-  
-    <script src="edit.js"></script>
-  </body>
-</html>
