+++ manifest.json
@@ -3,16 +3,16 @@
   "description": "Testing ",
   "version": "1.0",
   "manifest_version": 3,
-  "options_page": "options.html",
+  "options_page": "pages/options.html",
   "background": {
-    "service_worker": "background.js"
+    "service_worker": "./scripts/background.js"
   },
   "permissions": ["storage", "tabs"],
   "action": {
-    "default_popup": "popup.html",
+    "default_popup": "pages/popup.html",
     "default_icon": {
       "16": "/images/skeleton.png"
-  },
-  "icons": { "16": "/images/skeleton.png"
+    },
+    "icons": { "16": "/images/skeleton.png" }
   }
-  }}
+}
