+++ pages/newKeyForm.html
@@ -1,0 +1,53 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <link
+      href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@1,300&family=Roboto+Mono:ital,wght@1,300&family=Shippori+Antique&family=Shippori+Antique+B1&display=swap"
+      rel="stylesheet"
+    />
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
+      crossorigin="anonymous"
+    />
+    <link rel="stylesheet" href="../style.css" />
+  </head>
+  <body class="p-3 mb-2 bg-light text-dark wrapper">
+    <a href="./popup.html">
+      <img
+        src="../images/skeleton.png"
+        class="rounded-circle"
+        style="max-width: 9%"
+        alt="Skeleton Key Logo"
+      />
+    </a>
+
+    <h1 class="text-center p-1">New Key</h1>
+
+    <form id="newKeyForm" class="container">
+      <div class="mb-3">
+        <p class="form-label">Key Name</p>
+        <input
+          type="text"
+          class="form-control"
+          id="new_key_name"
+          placeholder="Holidays"
+        />
+      </div>
+
+      <button
+        type="button"
+        class="btn-sm btn-secondary fw-bolder"
+        id="submitKey"
+      >
+        Add Key
+      </button>
+    </form>
+
+    <script src="../scripts/newKeyForm.js"></script>
+  </body>
+</html>
