+++ pages/options.html
@@ -1,6 +1,5 @@
 <!DOCTYPE html>
-<a href="/popup.html"><img src="images/skeleton.png" class="rounded-circle" style="max-width:9%" alt="...">
-</a>
+
 <html>
   <head>
     <meta name="viewport" content="width=device-width, initial-scale=1"> </meta>
@@ -11,13 +10,20 @@
       integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
       crossorigin="anonymous"
     />
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="../style.css" />
   </head> 
+
    <body class="p-3 mb-2 bg-light text-dark wrapper">
-    <h1 class="text-center p-1">Skeleton Key Options</h1>
+    <a href="./popup.html">
+      <img src="../images/skeleton.png" class="rounded-circle" style="max-width:9%" alt="...">
+    </a>
+
+    <h1 class="text-center p-1">Options</h1>
+
     <div id="buttonDiv"></div>
+
     <a href="./newKeyForm.html" id="newKeyBtn" class="btn-sm btn-secondary fw-bolder">Create Key</a>
   
   </body>
-  <script type="module" src="options.js"></script>
+  <script type="module" src="../scripts/options.js"></script>
 </html>
