+++ pages/todolist.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <link rel="stylesheet" href="../style.css" />
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
+      crossorigin="anonymous"
+    />
+  </head>
+
+  <body class="wrapper2">
+    <a href="./popup.html">
+      <img
+        src="../images/skeleton.png"
+        class="rounded-circle"
+        style="max-width: 9%"
+        alt="..."
+      />
+    </a>
+    <div id="myDIV" class="header">
+      <h2 style="margin: 5px">My To Do List</h2>
+
+      <input type="text" id="myInput" placeholder="Go Shopping..." />
+
+      <span id="newtask" class="addBtn">Add</span>
+    </div>
+
+    <ul id="myUL"></ul>
+
+    <button id="deleteList" class="btn-sm btn-secondary fw-bolder">
+      Delete
+    </button>
+  </body>
+
+  <script type="module" src="../scripts/todolist.js"></script>
+</html>
