--- popup.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <a href="/options.html"><img src="images/skeleton.png" class="rounded-circle" style="max-width:9%" alt="...">
-    </a>
-      <meta charset="UTF-8" />
-      <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-      <link
-        href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-        rel="stylesheet"
-        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-        crossorigin="anonymous"
-      />
-    <link 
-      href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@1,300&family=Roboto+Mono:ital,wght@1,300&family=Shippori+Antique&family=Shippori+Antique+B1&display=swap"
-      rel="stylesheet"
-    />
-    <link rel="stylesheet" href="style.css" />
-  <h1 class="text-center p-1"> Skeleton Key </h1>
-  </head>
-  <body class="p-3 mb-2 bg-light text-dark wrapper">
-    <div>
-      <div id="buttonDiv" ></div>
-      <a href="options.html" class="btn btn-secondary btn-sm fw-bolder">Edit Keys</a>
-      <a href="todolist.html" class="btn btn-secondary btn-sm fw-bolder">To do list</a>
-    </div>
-    <script src="popup.js"></script>
-  </body>
-</html>
