+++ scripts/background.js
@@ -10,11 +10,12 @@ let keys = { work: workLinks, social: socialLinks };
 let todos = [];
 
 chrome.runtime.onInstalled.addListener(() => {
-  chrome.storage.sync.set({ keys });
+  chrome.storage.sync.set({ keys, todos });
   console.log(`work links defined`);
-});
-
-chrome.runtime.onInstalled.addListener(() => {
-  chrome.storage.sync.set({ todos });
   console.log(`todos created`);
 });
+
+// chrome.runtime.onInstalled.addListener(() => {
+//   chrome.storage.sync.set({ todos });
+
+// });
