+++ scripts/newKeyForm.js
@@ -14,5 +14,5 @@ function addKey(e) {
     chrome.storage.sync.set({ keys });
     document.forms[0].reset();
   });
-  window.location.href = "./options.html"
+  window.location.href = "../pages/options.html";
 }
